# Optimize encode32 and decode32