--- a/spec/ulid_spec.rb
+++ b/spec/ulid_spec.rb
@@ -26,7 +26,7 @@ describe ULID do
 
     it 'handles timestamp as the milliseconds precision' do
       expect(ULID.at(Time.parse('2016-07-30 22:36:16.001000000 UTC'))).to start_with('01ARYZ6RR1')
-      expect(ULID.at(Time.parse('2016-07-30 22:36:16.002000000 UTC'))).to start_with('01ARYZ6RR2')
+      expect(ULID.at(Time.parse('2016-07-30 22:36:16.002000001 UTC'))).to(start_with('01ARYZ6RR2')) # leap nanosecond
       expect(ULID.at(Time.parse('2016-07-30 22:36:16.003000000 UTC'))).to start_with('01ARYZ6RR3')
     end
 
@@ -46,7 +46,9 @@ describe ULID do
     it 'handles timestamp as the milliseconds precision' do
       time = ULID.time('0A000000000000000000000000')
       expect(time.to_i).to eq(10995116277)
-      expect(time.nsec).to eq(760000000)
+
+      # ULID precision is only to the millisecond value. Some parses will infer leap nanoseconds
+      expect(time.nsec).to be_within(1000).of(760000000)
     end
   end